+++ src/main/java/org/kumoricon/site/report/role/RoleReportView.java
@@ -30,17 +30,17 @@ public class RoleReportView extends BaseGridView implements View, ReportView {
 
     @PostConstruct
     public void init() {
-        setColumns(2);
+        setColumns(4);
         setRows(1);
         setColumnExpandRatio(0, 10);
-        setColumnExpandRatio(1, 1);
+        setColumnExpandRatio(3, 10);
         btnRefresh.addClickListener((Button.ClickListener) clickEvent -> handler.fetchReportData(this));
 
-        addComponent(data, 0, 0);
-        addComponent(btnRefresh, 1, 0);
+        addComponent(data, 1, 0);
+        addComponent(btnRefresh, 2, 0);
         data.setContentMode(ContentMode.HTML);
         handler.fetchReportData(this);
-        data.setWidth("100%");
+        data.setWidth("1000px");
     }
 
     public String getRequiredRight() {