+++ src/main/java/org/kumoricon/site/report/till/TillReportView.java
@@ -27,14 +27,14 @@ public class TillReportView extends BaseGridView implements View {
 
     @PostConstruct
     public void init() {
-        setColumns(2);
+        setColumns(4);
         setRows(1);
         setColumnExpandRatio(0, 10);
-        setColumnExpandRatio(1, 1);
-        addComponent(data, 0, 0);
-        addComponent(btnRefresh, 1, 0);
+        setColumnExpandRatio(3, 10);
+        addComponent(data, 1, 0);
+        addComponent(btnRefresh, 2, 0);
         data.setContentMode(ContentMode.HTML);
-        data.setWidth("100%");
+        data.setWidth("1000px");
         data.setHeightUndefined();
 
         btnRefresh.addClickListener((Button.ClickListener) clickEvent -> {